requirements.txt
flask>=3.0.0
flask-jwt-extended>=4.6.0
flask-cors>=4.0.0
pandas>=2.1.0
openpyxl>=3.1.0
xlrd>=2.0.1
lxml>=5.0.0
supabase>=2.28.2
python-dotenv>=1.0.1
gunicorn>=21.2.0
